+++ src/main/kotlin/com/expedia/graphql/schema/SchemaGeneratorConfig.kt
@@ -1,5 +1,6 @@
 package com.expedia.graphql.schema
 
+import com.expedia.graphql.schema.generator.completableFutureResolver
 import com.expedia.graphql.schema.hooks.NoopSchemaGeneratorHooks
 import com.expedia.graphql.schema.hooks.SchemaGeneratorHooks
 import kotlin.reflect.KType
